+++ packages/asc-ui/src/components/Heading/HeadingStyle.ts
@@ -6,8 +6,6 @@ export type Props = TypographyProps
 
 export const HeaderStyleCSS = css`
   margin-top: 0;
-
-  color: green;
 `
 
 export default styled(Typography)<Props>`