+++ packages/asc-ui/src/theme/default/spacing.ts
@@ -1,5 +1,5 @@
 import type { Theme } from '../../types'
 
-const spacing: Theme.Spacing = 4
+const spacing: Theme.Spacing = 8
 
 export default spacing