+++ packages/asc-ui/src/utils/index.ts
@@ -15,6 +15,7 @@ export {
   CustomCssPropsInterface,
   getValueFromTheme,
   themeSpacing,
+  calculateFluidStyle,
 } from './themeUtils'
 export { valueFromObject } from './valueFromObject'
 export { fromProps } from './fromProps'